Rename TypeManager to DynamicTupleCache
