+++ Cql/CodeGeneration.NET/AssemblyCompiler.cs
@@ -28,7 +28,7 @@ namespace Hl7.Cql.CodeGeneration.NET
 {
     internal class AssemblyCompiler
     {
-        private readonly TypeManager _typeManager;
+        private readonly DynamicTupleCache _dynamicTupleCache;
         private readonly CSharpLibrarySetToStreamsWriter _cSharpLibrarySetToStreamsWriter;
         private readonly CSharpCodeStreamPostProcessor? _cSharpCodeStreamPostProcessor;
         private readonly Lazy<Assembly[]> _referencesLazy;
@@ -36,13 +36,13 @@ namespace Hl7.Cql.CodeGeneration.NET
 
         public AssemblyCompiler(
             CSharpLibrarySetToStreamsWriter cSharpLibrarySetToStreamsWriter,
-            TypeManager typeManager,
+            DynamicTupleCache dynamicTupleCache,
             CSharpCodeStreamPostProcessor? cSharpCodeStreamPostProcessor = null,
             AssemblyDataPostProcessor? assemblyDataPostProcessor = null)
         {
             _assemblyDataPostProcessor = assemblyDataPostProcessor;
             _cSharpLibrarySetToStreamsWriter = cSharpLibrarySetToStreamsWriter;
-            _typeManager = typeManager;
+            _dynamicTupleCache = dynamicTupleCache;
             _cSharpCodeStreamPostProcessor = cSharpCodeStreamPostProcessor;
             _referencesLazy = new Lazy<Assembly[]>(
                 () =>
@@ -62,7 +62,7 @@ namespace Hl7.Cql.CodeGeneration.NET
 
                         }                                        // @formatter on
                         .Select(type => type.Assembly)
-                        .Concat(typeManager.Resolver.ModelAssemblies)
+                        .Concat(dynamicTupleCache.Resolver.ModelAssemblies)
                         .Distinct()
                         .ToArray();
                     return references;
